Move validations to validations.ts for use in web worker
