+++ app/client/src/constants/WidgetValidation.ts
@@ -26,6 +26,7 @@ export const VALIDATION_TYPES = {
   DEFAULT_SELECTED_ROW: "DEFAULT_SELECTED_ROW",
   COLUMN_PROPERTIES_ARRAY: "COLUMN_PROPERTIES_ARRAY",
   LAT_LONG: "LAT_LONG",
+  TABLE_PAGE_NO: "TABLE_PAGE_NO",
 };
 
 export type ValidationResponse = {
