+++ app/client/src/workers/validations.test.ts
@@ -17,9 +17,9 @@ const DUMMY_WIDGET: WidgetProps = {
   widgetName: "",
 };
 
-describe("Chart data validator", () => {
+describe("Validate Validators", () => {
   const validator = VALIDATORS.CHART_DATA;
-  it("correctly validates ", () => {
+  it("correctly validates chart data ", () => {
     const cases = [
       {
         input: [
@@ -118,5 +118,24 @@ describe("Chart data validator", () => {
       const response = validator(testCase.input, DUMMY_WIDGET, {});
       expect(response).toStrictEqual(testCase.output);
     }
+  });
+  it("Correctly validates page number", () => {
+    const input = [0, -1, undefined, null, 2, "abcd", [], ""];
+    const expected = [1, 1, 1, 1, 2, 1, 1, 1];
+    input.forEach((val, index) => {
+      const result = VALIDATORS.TABLE_PAGE_NO(val, DUMMY_WIDGET, undefined);
+      const expectedResult: {
+        isValid: boolean;
+        parsed: number;
+        message?: string;
+      } = {
+        isValid: expected[index] !== 1,
+        parsed: expected[index],
+      };
+      if (expected[index] === 1) {
+        expectedResult.message = "";
+      }
+      expect(result).toStrictEqual(expectedResult);
+    });
   });
 });
